- Ajustes e debug em console

+++ Transactions/SyncronizationBot.Domain/Model/Database/RunTimeController.cs
@@ -6,7 +6,7 @@ namespace SyncronizationBot.Domain.Model.Database
     public class RunTimeController : Entity
     {
         public int? IdRuntime { get; set; }
-        public int? ConfigurationTimer { get; set; }
+        public decimal? ConfigurationTimer { get; set; }
         public ETypeService TypeService { get; set; }
         public bool? IsRunning { get; set; }
         public bool? IsContingecyTransactions { get; set; }
